Task 4 implement fetch simulation method
Implemented the fetch() simulation member method to
perform the fetch part of the fetch/execute cycle.
The fetch stage is simple, whatever address that the pc is
currently pointing to, we fetch that value from memory and
store it into the ir instruction register, in preparation
for decoding and execution.  All task 4 unit tests
are passing for this commit.