Task 6 implement executeX() member functions
All implementations of the separate member functions
to execute the simulated machine instructions have been
added in this commit.  Added in basic switch statement to
invoke indicated method to perform execution of the
current decoded instruction.  All task 6 tests are passing
for the set of execute methods used in this simulation.